+++ src/main/java/com/hrms/entity/MenuRole.java
@@ -9,10 +9,10 @@ import lombok.Data;
 @Data
 public class MenuRole implements Serializable{
 
-  private Integer id; //主键id
+  private Long id; //主键id
 
-  private Integer mid; //菜单id
+  private Long mid; //菜单id
 
-  private Integer rid; //角色id
+  private Long rid; //角色id
 
 }
